+++ src/routers/invoice.router.ts
@@ -24,3 +24,24 @@ invoiceRouter.get(
     }
   }
 );
+invoiceRouter.get(
+  "/:id",
+  param("id").exists().isNumeric(),
+  async (req: Request, res: Response) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
+    }
+    try {
+      const id: number = await parseInt(req.params.id, 10);
+      const invoiceDetailList = await invoiceV2Service.getInvoiceDetail(id);
+      if (invoiceDetailList != undefined) {
+        return res.status(200).json(invoiceDetailList);
+      } else {
+        return res.status(404).send(errorHandling(undefined));
+      }
+    } catch (err) {
+      res.status(401).json(err);
+    }
+  }
+);
