+++ src/services/invoiceV2.service.ts
@@ -8,4 +8,7 @@ export class InvoiceV2Service {
   public async getAllInvoicesPaginated(pageNumber: number) {
     return await this.invoiceRepository.getInvoicesPaginated(pageNumber);
   }
+  public async getInvoiceDetail(invoiceId: number) {
+    return await this.invoiceRepository.getInvoiceDetail(invoiceId);
+  }
 }
